# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Content Workflow API
  description: |
    REST API for AI-powered content creation, translation, and review workflows.
    
    This API provides comprehensive endpoints for managing marketing campaigns,
    content pieces, AI generation, translations, and review processes.
    
    **Features:**
    - Campaign and content management
    - Multi-model AI content generation (OpenAI, Anthropic)
    - Automated translation services
    - Content analysis and sentiment detection
    - Review and approval workflows
    - Real-time monitoring and health checks
    
  version: 1.0.0
  contact:
    name: AI Content Workflow Team
    url: https://github.com/your-org/ai-content-workflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server
paths:
...
  /content/{id}/translate:
    post:
      summary: Translate content
      description: Create a translation of the content piece
      operationId: translateContent
      tags:
        - Translation
      parameters:
        - name: id
          in: path
          required: true
          description: Content piece ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
      responses:
        '200':
          description: Translation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Translation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Translation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        targetLanguage:
          type: string
          description: Target language code
          example: "es"
        translatedText:
          type: string
          description: Translated content
        status:
          type: string
          enum: [draft, completed]
        qualityScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
          description: Translation quality score
        createdAt:
          type: string
          format: date-time
...
    TranslateRequest:
      type: object
      required:
        - targetLanguage
      properties:
        targetLanguage:
          type: string
          example: "es"
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request data"
            code: "BAD_REQUEST"
            timestamp: "2024-01-01T12:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            timestamp: "2024-01-01T12:00:00Z"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error occurred"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-01T12:00:00Z"
...